create separate packageJSON helper in CJS to avoid breaking esbuild
